Fix Rust 1.66 clippy lints including `clippy::unused_format_specs`
`clippy::unused_format_specs` pointed out that we were actually
concatenating `device.arch()` to `device.details()` because alignment
args don't apply to nested `format_args!` invocations.

For example, see how the last two columns are concatenated instead of
aligned (alignment shortened for brevity elsewhere):

    host            Linux   linux x64Arch Linux 6.0.12-arch1-1
    adb:mydeviceid  b0s     android arm64Android 13 (API 33)

+++ xcommon/src/lib.rs
@@ -252,7 +252,7 @@ fn find_cde_start_pos<R: Read + Seek>(reader: &mut R) -> Result<u64> {
     anyhow::ensure!(file_length >= HEADER_SIZE, "Invalid zip header");
     let mut pos = file_length - HEADER_SIZE;
     while pos >= search_upper_bound {
-        reader.seek(SeekFrom::Start(pos as u64))?;
+        reader.seek(SeekFrom::Start(pos))?;
         if reader.read_u32::<LittleEndian>()? == CENTRAL_DIRECTORY_END_SIGNATURE {
             return Ok(pos);
         }
